Update Roadmap in realtime
cut development time down even more and have included many iterations of testing to get to this final phase of the project. This file will probably end up in a docs directory after a final release is built. I may include a donation address in the about section where all sources and credit due will be listed. Thanks to all those before me who inspired me with their works to create this beast of a full-node-utility-miner for cross-bitcoin chain mining and testing. I may include a second release for pools and exchanges especially after the atomic swapping is integrated. This would allow users to bypass the need for exchanges as everything would all be in one place and would allow users to store their own keys.